fix register-middleware
