+++ src/FilamentLanguageSwitchServiceProvider.php
@@ -2,6 +2,7 @@
 
 namespace BezhanSalleh\FilamentLanguageSwitch;
 
+use BezhanSalleh\FilamentLanguageSwitch\Http\Middleware\SwitchLanguageLocale;
 use Filament\Facades\Filament;
 use Filament\PluginServiceProvider;
 use Spatie\LaravelPackageTools\Package;
@@ -24,6 +25,18 @@ class FilamentLanguageSwitchServiceProvider extends PluginServiceProvider
 
     public function packageBooted(): void
     {
+        $this->registerSwitchLanguageMiddleware();
         Filament::serving(fn () => FilamentLanguageSwitch::boot());
+    }
+
+    public function registerSwitchLanguageMiddleware(): void
+    {
+        if (! array_key_exists(
+            $key = SwitchLanguageLocale::class,
+            $filamentMiddlewares = config('filament.middleware.base')
+        )) {
+            $filamentMiddlewares[] = $key;
+            config(['filament.middleware.base' => $filamentMiddlewares]);
+        }
     }
 }
